+++ src/App.tsx
@@ -1,15 +1,13 @@
 import { Routes, Route } from "react-router-dom";
-import LoginPage from "./pages/LoginPgae/LoginPage";
-import SignUpPage from "./pages/SignUpPage/SignUpPage";
+import AuthPage from "./pages/AuthPage/AuthPage";
 
 
 const App = () => {
 
   return (
     <Routes>
-    <Route path="/" element={<LoginPage />} />
-    <Route path="/signUp" element={<SignUpPage />} />
-</Routes>
+    <Route path="/" element={<AuthPage />} />
+    </Routes>
   );
 };
