+++ src/pages/AuthPage/authPage.scss
@@ -1,0 +1,66 @@
+@import url('https://fonts.googleapis.com/css2?family=Varela+Round&display=swap');
+
+$font-main: 'Varela Round', sans-serif;
+$color-primary: #0751D8;
+
+html,
+body {
+  text-decoration: none;
+  margin: 0;
+  padding: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background-color: #f5f5f5;
+}
+
+.auth-page {
+  flex-direction: column;
+  padding: 20px 30px;
+  align-items: center;
+  justify-content: flex-start;
+  width: 400px;
+  background: white;
+  box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
+  border-radius: 10px;
+
+  h5 {
+    font-size: 24px;
+    font-weight: 500;
+    font-family: $font-main;
+    text-align: center;
+    margin-bottom: 15px;
+  }
+}
+
+.auth-form {
+  width: 100%;
+
+  .input-field {
+    margin-top: 10px;
+
+    .MuiInputBase-input {
+      font-size: 14px;
+      font-family: $font-main;
+    }
+  }
+
+  .button-submit {
+    margin-top: 15px;
+    box-shadow: none;
+    background: $color-primary;
+    border-radius: 8px;
+    font-family: $font-main;
+    color: white;
+    text-transform: none;
+  }
+}
+
+.MuiButton-root {
+  text-transform: none;
+}
+.auth-page .MuiButton-root {
+    text-decoration: none; 
+  }
