+++ src/shared/api/auth/AuthService.ts
@@ -1,4 +1,4 @@
-import { supabase } from "../../utils/client";
+import { supabase } from "../../../utils/client";
 
 export const AuthService = {
     async login(email: string, password: string) {
